% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_crops.R
\name{post_crop_table}
\alias{post_crop_table}
\title{post_crop_table}
\usage{
post_crop_table(table_crops)
}
\arguments{
\item{table_crops}{a data frame}
}
\description{
store crop table
}
\author{
Reinhard Simon
}

