% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/server_crop.R
\name{server_crop}
\alias{server_crop}
\title{Server side logic}
\usage{
server_crop(input, output, session, dom = "hot_crops")
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{dom}{target dom element name}
}
\description{
Constructs table
}
\author{
Reinhard Simon
}

